+++ plugins/database/mssql/mssql.go
@@ -12,8 +12,8 @@ import (
 	"github.com/hashicorp/errwrap"
 	"github.com/hashicorp/vault/api"
 	"github.com/hashicorp/vault/builtin/logical/database/dbplugin"
+	"github.com/hashicorp/vault/helper/dbtxn"
 	"github.com/hashicorp/vault/helper/strutil"
-	"github.com/hashicorp/vault/helper/transaction"
 	"github.com/hashicorp/vault/plugins"
 	"github.com/hashicorp/vault/plugins/helper/database/connutil"
 	"github.com/hashicorp/vault/plugins/helper/database/credsutil"
@@ -130,13 +130,13 @@ func (m *MSSQL) CreateUser(ctx context.Context, statements dbplugin.Statements, 
 				continue
 			}
 
-			c := &transaction.Config{
+			c := &dbtxn.Config{
 				Name:       username,
 				Password:   password,
 				Expiration: expirationStr,
 			}
 
-			if err := transaction.ExecuteTxQuery(ctx, tx, c, query); err != nil {
+			if err := dbtxn.ExecuteTxQuery(ctx, tx, c, query); err != nil {
 				return "", "", err
 			}
 		}
@@ -187,10 +187,10 @@ func (m *MSSQL) RevokeUser(ctx context.Context, statements dbplugin.Statements, 
 				continue
 			}
 
-			c := &transaction.Config{
+			c := &dbtxn.Config{
 				Name: username,
 			}
-			if err := transaction.ExecuteTxQuery(ctx, tx, c, query); err != nil {
+			if err := dbtxn.ExecuteTxQuery(ctx, tx, c, query); err != nil {
 				return err
 			}
 		}
@@ -279,7 +279,7 @@ func (m *MSSQL) revokeUserDefault(ctx context.Context, username string) error {
 	// many permissions as possible right now
 	var lastStmtError error
 	for _, query := range revokeStmts {
-		if err := transaction.ExecuteDBQuery(ctx, db, nil, query); err != nil {
+		if err := dbtxn.ExecuteDBQuery(ctx, db, nil, query); err != nil {
 			lastStmtError = err
 		}
 	}
@@ -343,11 +343,11 @@ func (m *MSSQL) RotateRootCredentials(ctx context.Context, statements []string) 
 				continue
 			}
 
-			c := &transaction.Config{
+			c := &dbtxn.Config{
 				Username: m.Username,
 				Password: password,
 			}
-			if err := transaction.ExecuteTxQuery(ctx, tx, c, query); err != nil {
+			if err := dbtxn.ExecuteTxQuery(ctx, tx, c, query); err != nil {
 				return nil, err
 			}
 		}
