+++ plugins/database/mysql/mysql.go
@@ -10,8 +10,8 @@ import (
 	stdmysql "github.com/go-sql-driver/mysql"
 	"github.com/hashicorp/vault/api"
 	"github.com/hashicorp/vault/builtin/logical/database/dbplugin"
+	"github.com/hashicorp/vault/helper/dbtxn"
 	"github.com/hashicorp/vault/helper/strutil"
-	"github.com/hashicorp/vault/helper/transaction"
 	"github.com/hashicorp/vault/plugins"
 	"github.com/hashicorp/vault/plugins/helper/database/connutil"
 	"github.com/hashicorp/vault/plugins/helper/database/credsutil"
@@ -303,12 +303,12 @@ func (m *MySQL) RotateRootCredentials(ctx context.Context, statements []string) 
 				continue
 			}
 
-			c := &transaction.Config{
+			c := &dbtxn.Config{
 				Username: m.Username,
 				Password: password,
 			}
 
-			if err := transaction.ExecuteTxQuery(ctx, tx, c, query); err != nil {
+			if err := dbtxn.ExecuteTxQuery(ctx, tx, c, query); err != nil {
 				return nil, err
 			}
 		}
